more work on fluid kernels and host side, a bunch of bug fixes

+++ physics/ocl/kernels/grid.cl
@@ -15,7 +15,7 @@ inline int3 getCell3DIndexFromPos(float4 pos)
   // Moving particles in [0 - 2 * ABS_WALL_POS] to have coords matching with cellIndices
   const float3 posXYZ = clamp(pos.xyz, -ABS_WALL_POS, ABS_WALL_POS) + (float3)(ABS_WALL_POS, ABS_WALL_POS, ABS_WALL_POS);
 
-  const int cellSize = 2 * ABS_WALL_POS / GRID_RES;
+  const float cellSize = 2 * ABS_WALL_POS / GRID_RES;
 
   const int3 cell3DIndex = convert_int3(posXYZ / cellSize);
